requirements.txt
pandas
numpy
matplotlib
python-binance
ta-lib
configparser
